requirements.txt
streamlit
torch
sentence-transformers
transformers
python-dotenv
pinecone-client